Add files via upload
Overview:
This repository contains the projects completed during my internship at Academor. It includes a minor project focused on Indian historical places and a major project on a weather forecasting application. These projects provided hands-on experience in web development using HTML, CSS, and JavaScript.
Projects
1)Minor Project: Indian Historical Places
Description
An interactive web application that showcases various historical places in India. The aim is to educate users about the significance, history, and key features of these sites through a visually appealing and user-friendly interface.

Technologies Used
HTML: For structuring the web pages.
CSS: For styling and enhancing the visual presentation.
JavaScript: For adding interactivity and dynamic content.
2)Major Project: Weather Forecasting App
Description
A web application that provides a 7-day weather forecast, including temperature, wind speed, and humidity information. The app aims to deliver accurate and up-to-date weather information in an easy-to-use format.

Technologies Used
HTML: For structuring the application.
CSS: For styling and ensuring a consistent look and feel.
JavaScript: For fetching and displaying dynamic weather data.